Use explicit convertion from int/long to double for abs

+++ src/settings.cpp
@@ -1467,7 +1467,7 @@ bool Settings::PerspectiveExists(const wxString& perspective_name)
 
 void Settings::SetAutoloadedChatlogLinesCount(const int count)
 {
-	cfg().Write(_T( "/GUI/AutoloadedChatlogLinesCount" ), std::abs(count));
+	cfg().Write(_T( "/GUI/AutoloadedChatlogLinesCount" ), std::abs(static_cast<double>(count)));
 }
 
 int Settings::GetAutoloadedChatlogLinesCount()
